requirements.txt
streamlit
pandas
matplotlib
openpyxl
Pillow